feat: Program 이미지 추가(입력과 수정 테스트 필요)

+++ src/main/java/com/unitalk/common/model/entity/Student.java
@@ -24,7 +24,7 @@ public class Student {
     private String name; // 이름
 
     @Column(nullable = false)
-    private Integer grade; // 학년
+    private Long grade; // 학년
 
     @Column(nullable = false)
     private LocalDateTime admissionDate; // 입학일
