+++ src/main/java/com/unitalk/program/model/dto/response/ProgramImgResponse.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class ProgramImgResponse {
-    private Integer imgId; // 집단상담 파일 번호(PK)
+    private Long imgId; // 집단상담 파일 번호(PK)
     private Program programId; // 집단상담 번호(FK)
     private String imgName; // 집단상담 파일명
     private String imgSaveName; // 집단상담 파일 저장명
