+++ src/main/java/com/unitalk/program/repository/ProgramImgRepository.java
@@ -3,5 +3,5 @@ package com.unitalk.program.repository;
 import com.unitalk.program.model.entity.ProgramImg;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProgramImgRepository extends JpaRepository<ProgramImg, Integer> {
+public interface ProgramImgRepository extends JpaRepository<ProgramImg, Long> {
 }
