+++ src/main/java/com/unitalk/program/repository/ProgramResultRepository.java
@@ -3,5 +3,5 @@ package com.unitalk.program.repository;
 import com.unitalk.program.model.entity.ProgramResult;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProgramResultRepository extends JpaRepository<ProgramResult, Integer> {
+public interface ProgramResultRepository extends JpaRepository<ProgramResult, Long> {
 }
