+++ src/main/java/com/unitalk/program/repository/ProgramReviewRepository.java
@@ -3,5 +3,5 @@ package com.unitalk.program.repository;
 import com.unitalk.program.model.entity.ProgramReview;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProgramReviewRepository extends JpaRepository<ProgramReview, Integer> {
+public interface ProgramReviewRepository extends JpaRepository<ProgramReview, Long> {
 }
